Fix ruby 2.8 unpacking data
When unpacking the data stream using ruby 2.8 the unpacking process will sometimes randomly pick the default choice.

Removing the default choice reveals there is an issue with the selection where BinData claims the selection doesn't exist.

IndexError:
  selection '1' does not exist in :choices for obj.val

I have found that explicitly calling to_i when on the type_id of the selection fixes the issue.